move definitions from head files
