+++ compile_run.bat
@@ -1,3 +1,3 @@
-gcc .\lib\hashids\hashids.c .\src\tests.c -o ggeny.exe
+gcc src\ggeny.c src\ggenyrand.c src\ggenygraph.c src\ggenyio.c lib\hashids\hashids.c -o ggeny.exe
 ggeny.exe
 pause
